third_party/blink/web_tests/external/wpt/domparsing/DOMParser-parseFromString-xml-expected.txt

This is a testharness.js-based test.
[FAIL] XMLDocument interface for correctly parsed document with type text/xml
  assert_false: Should not be XMLDocument expected false got true
[FAIL] Should return an error document for XML wellformedness errors in type text/xml
  assert_equals: expected (string) "http://www.mozilla.org/newlayout/xml/parsererror.xml" but got (object) null
[FAIL] XMLDocument interface for incorrectly parsed document with type text/xml
  assert_false: Should not be XMLDocument expected false got true
[FAIL] XMLDocument interface for correctly parsed document with type application/xml
  assert_false: Should not be XMLDocument expected false got true
[FAIL] Should return an error document for XML wellformedness errors in type application/xml
  assert_equals: expected (string) "http://www.mozilla.org/newlayout/xml/parsererror.xml" but got (object) null
[FAIL] XMLDocument interface for incorrectly parsed document with type application/xml
  assert_false: Should not be XMLDocument expected false got true
[FAIL] XMLDocument interface for correctly parsed document with type application/xhtml+xml
  assert_false: Should not be XMLDocument expected false got true
[FAIL] Should return an error document for XML wellformedness errors in type application/xhtml+xml
  assert_equals: expected (string) "http://www.mozilla.org/newlayout/xml/parsererror.xml" but got (object) null
[FAIL] XMLDocument interface for incorrectly parsed document with type application/xhtml+xml
  assert_false: Should not be XMLDocument expected false got true
[FAIL] XMLDocument interface for correctly parsed document with type image/svg+xml
  assert_false: Should not be XMLDocument expected false got true
[FAIL] Should return an error document for XML wellformedness errors in type image/svg+xml
  assert_equals: expected (string) "http://www.mozilla.org/newlayout/xml/parsererror.xml" but got (object) null
[FAIL] XMLDocument interface for incorrectly parsed document with type image/svg+xml
  assert_false: Should not be XMLDocument expected false got true
Harness: the test ran to completion.
